Move Job class into new namespace
