+++ lib/bulk_processor.rb
@@ -2,7 +2,7 @@ require 'bulk_processor/back_end'
 require 'bulk_processor/back_end/active_job'
 require 'bulk_processor/back_end/dynosaur'
 require 'bulk_processor/config'
-require 'bulk_processor/job'
+require 'bulk_processor/job/process_csv'
 require 'bulk_processor/payload_serializer'
 require 'bulk_processor/s3_file'
 require 'bulk_processor/stream_encoder'
