--- lib/bulk_processor/job.rb
@@ -1,20 +1,0 @@
-require 'active_job'
-
-class BulkProcessor
-  # ActiveJob to handle processing the CSV in the background
-  class Job < ActiveJob::Base
-    queue_as 'bulk_processor'
-
-    def perform(processor_class, payload, key)
-      file = BulkProcessor.config.file_class.new(key)
-      payload = PayloadSerializer.deserialize(payload)
-      file.open do |f|
-        csv = CSV.parse(f.read, headers: true)
-        processor = processor_class.constantize.new(csv, payload: payload)
-        processor.start
-      end
-    ensure
-      file.try(:delete)
-    end
-  end
-end
