+++ spec/bulk_processor/tasks_spec.rb
@@ -3,11 +3,11 @@ require 'bulk_processor/tasks'
 
 describe 'bulk_processor namespace' do
   describe 'bulk_processor:start' do
-    let(:job) { instance_double(BulkProcessor::Job, perform: true) }
+    let(:job) { instance_double(BulkProcessor::Job::ProcessCSV, perform: true) }
 
     before(:all) { Rake::Task.define_task(:environment) }
 
-    before { allow(BulkProcessor::Job).to receive(:new).and_return(job) }
+    before { allow(BulkProcessor::Job::ProcessCSV).to receive(:new).and_return(job) }
 
     it 'starts the job' do
       expect(job).to receive(:perform)
