+++ spec/support/role/csv_processor.rb
@@ -1,6 +1,6 @@
 class BulkProcessor
   module Role
-    # Role used by BulkProcessor::Job to process a set of records from a CSV.
+    # Role used by BulkProcessor::Job::ProcessCSV to process a set of records from a CSV.
     class CSVProcessor
       def self.required_columns
       end
